+++ packages/categorize/configure/src/design/categories/choice-preview.jsx
@@ -1,18 +1,28 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import withStyles from '@mui/styles/withStyles';
-import classNames from 'classnames';
-import { Choice } from '@pie-lib/drag';
+import { styled } from '@mui/material/styles';
+import { DraggableChoice } from '@pie-lib/drag';
 import IconButton from '@mui/material/IconButton';
 import RemoveCircleOutlineIcon from '@mui/icons-material/RemoveCircleOutline';
 import { HtmlAndMath } from '@pie-lib/render-ui';
 import { color } from '@pie-lib/render-ui';
 
+const ChoicePreviewContainer = styled('div')({
+  position: 'relative',
+  overflow: 'auto',
+});
+
+const DeleteIconButton = styled(IconButton)({
+  position: 'absolute',
+  right: 0,
+  top: 0,
+  color: `${color.tertiary()} !important`,
+});
+
 export class ChoicePreview extends React.Component {
   static propTypes = {
     alternateResponseIndex: PropTypes.number,
     category: PropTypes.object,
-    classes: PropTypes.object.isRequired,
     className: PropTypes.string,
     choice: PropTypes.object.isRequired,
     choiceIndex: PropTypes.number,
@@ -26,51 +36,32 @@ export class ChoicePreview extends React.Component {
   };
 
   render() {
-    const { alternateResponseIndex, category, classes, className, choice, choiceIndex } = this.props;
+    const { alternateResponseIndex, category, className, choice, choiceIndex } = this.props;
     return (
-      <div className={classNames(classes.choicePreview, className)}>
+      <ChoicePreviewContainer className={className}>
         {choice ? (
-          <Choice
+          <DraggableChoice
             alternateResponseIndex={alternateResponseIndex}
             category={category}
             choice={choice}
             choiceIndex={choiceIndex}
-            className={classes.overflowChoice}
-            onRemoveChoice={() => this.delete()}
+            onRemoveChoice={this.delete}
+            type={'choice-preview'}
+            id={choice.id}
+            categoryId={category && category.id}
           >
-            <HtmlAndMath html={choice?.content} className={`${classes.breakWord}`} />
-            <IconButton
-              aria-label="delete"
-              className={classNames(classes.delete, classes.customColor)}
-              onClick={this.delete}
-              size="large">
-              <RemoveCircleOutlineIcon />
-            </IconButton>
-          </Choice>
+            <HtmlAndMath html={choice?.content} />
+          </DraggableChoice>
         ) : null}
-      </div>
+        <DeleteIconButton
+          aria-label="delete"
+          onClick={this.delete}
+          size="large">
+          <RemoveCircleOutlineIcon />
+        </DeleteIconButton>
+      </ChoicePreviewContainer>
     );
   }
 }
-const styles = () => ({
-  choicePreview: {
-    position: 'relative',
-    overflow: 'auto',
-  },
-  delete: {
-    position: 'absolute',
-    right: 0,
-    top: 0,
-  },
-  breakWord: {
-    maxWidth: '90%',
-    wordBreak: 'break-all',
-  },
-  customColor: {
-    color: `${color.tertiary()} !important`,
-  },
-  overflowChoice: {
-    overflow: 'auto',
-  },
-});
-export default withStyles(styles)(ChoicePreview);
+
+export default ChoicePreview;