+++ packages/multiple-choice/src/multiple-choice.jsx
@@ -115,7 +115,7 @@ export class MultipleChoice extends React.Component {
     super(props);
 
     this.state = {
-      showCorrect: this.props.alwaysShowCorrect || false,
+      showCorrect: (this.props.options && this.props.alwaysShowCorrect) || false,
       maxSelectionsErrorState: false,
     };
 
@@ -166,7 +166,7 @@ export class MultipleChoice extends React.Component {
       });
     }
 
-    if (nextProps.alwaysShowCorrect && this.state.showCorrect !== true) {
+    if (nextProps.options && nextProps.alwaysShowCorrect && this.state.showCorrect !== true) {
       this.setState({ showCorrect: true }, () => {
         if (this.props.onShowCorrectToggle) {
           this.props.onShowCorrectToggle();
@@ -285,6 +285,7 @@ export class MultipleChoice extends React.Component {
       autoplayAudioEnabled,
       session,
       customAudioButton,
+      options,
     } = this.props;
     const { showCorrect, maxSelectionsErrorState } = this.state;
     const isEvaluateMode = mode === 'evaluate';
@@ -362,7 +363,7 @@ export class MultipleChoice extends React.Component {
             customAudioButton={customAudioButton}
           />
 
-          {!alwaysShowCorrect && (
+          {!(options && alwaysShowCorrect) && (
             <CorrectAnswerToggle
               show={showCorrectAnswerToggle}
               toggled={showCorrect}