+++ packages/matrix/controller/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-controller@1.1.1...@pie-element/matrix-controller@1.2.1) (2025-10-16)
+
+**Note:** Version bump only for package @pie-element/matrix-controller
+
+
+
+
+
+# [1.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-controller@1.1.1...@pie-element/matrix-controller@1.2.0) (2025-10-15)
+
+**Note:** Version bump only for package @pie-element/matrix-controller
+
+
+
+
+
 ## [1.1.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-controller@1.0.1...@pie-element/matrix-controller@1.1.1) (2025-10-10)
 
 **Note:** Version bump only for package @pie-element/matrix-controller